--- a/.github/workflows/ci.yaml
+++ b/.github/workflows/ci.yaml
@@ -83,6 +83,8 @@ jobs:
         uses: actions/setup-python@v2
         with:
           python-version: ${{ matrix.python-version }}
+      - name: Install uv
+        uses: astral-sh/setup-uv@v5
       - name: Set up pip cache
         uses: actions/cache@v3
         with:
@@ -90,9 +92,9 @@ jobs:
           key: ${{ matrix.tox-env }}-${{ matrix.os }}
       - name: Set up environment
         run: |
-          pip install --upgrade pip
-          pip install mypy==1.15.0
-          pip install ".[dev]"
+          uv pip install --upgrade pip
+          uv pip install mypy==1.15.0
+          uv pip install ".[dev]"
       - name: Run alembic
         run: |
           export PYTHONPATH=$(pwd)/src:$PYTHONPATH