--- a/.github/workflows/ci.yaml
+++ b/.github/workflows/ci.yaml
@@ -85,16 +85,15 @@ jobs:
           python-version: ${{ matrix.python-version }}
       - name: Install uv
         uses: astral-sh/setup-uv@v5
-      - name: Set up pip cache
-        uses: actions/cache@v3
-        with:
-          path: .cache/pip
-          key: ${{ matrix.tox-env }}-${{ matrix.os }}
+      - name: Create virtualenv
+        run: uv venv
+      - name: Install dependencies
+        run: uv pip sync
       - name: Set up environment
         run: |
           uv pip install --upgrade pip
           uv pip install mypy==1.15.0
-          uv pip install ".[dev]"
+          uv pip sync
       - name: Run alembic
         run: |
           export PYTHONPATH=$(pwd)/src:$PYTHONPATH
@@ -103,5 +102,5 @@ jobs:
         run: |
           # Include src/ directory in Python path to prioritize local files in pytest
           export PYTHONPATH=$(pwd)/src:$PYTHONPATH
-          mypy src/
-          pytest --color=yes tests/
+          uv run mypy src/
+          uv run pytest --color=yes tests/