--- a/.github/workflows/ci.yaml
+++ b/.github/workflows/ci.yaml
@@ -30,20 +30,16 @@ jobs:
         uses: actions/setup-python@v2
         with:
           python-version: "3.11"
-      - name: Set up pip cache
-        uses: actions/cache@v3
+      - name: Install uv
+        uses: astral-sh/setup-uv@v5
         with:
-          path: ~/.cache/pip
-          key: ${{ runner.os }}-pip-${{ hashFiles('**/pyproject.toml') }}
-      - name: Set up environment
-        run: |
-          pip install --upgrade pip wheel setuptools
-          pip install bandit[toml]==1.7.4 ruff==0.6.7
+          enable-cache: true
+          cache-dependency-glob: "backend/uv.lock"
       - name: Linting check
         run: |
-          bandit -qr -c pyproject.toml src/
-          ruff check src/ tests/
-          ruff format --check src/ tests/
+          uv run bandit -qr -c pyproject.toml src/
+          uv run ruff check src/ tests/
+          uv run ruff format --check src/ tests/
 
   mypy:
     runs-on: ${{ matrix.os }}