--- a/.github/workflows/ci.yaml
+++ b/.github/workflows/ci.yaml
@@ -35,6 +35,8 @@ jobs:
         with:
           enable-cache: true
           cache-dependency-glob: "backend/uv.lock"
+      - name: create venv
+        run: uv venv
       - name: Set up environment
         run: uv pip install bandit[toml] ruff
       - name: Linting check
@@ -72,7 +74,7 @@ jobs:
           enable-cache: true
           cache-dependency-glob: "backend/uv.lock"
       - name: Install the project
-        run: uv sync --locked --all-extras --dev
+        run: uv sync --locked --all-extras
       - name: Run mypy
         run: |
           export PYTHONPATH=$(pwd)/src:$PYTHONPATH
@@ -120,7 +122,7 @@ jobs:
           enable-cache: true
           cache-dependency-glob: "backend/uv.lock"
       - name: Install the project
-        run: uv sync --locked --all-extras --dev
+        run: uv sync --locked --all-extras
       - name: Run alembic
         run: |
           export PYTHONPATH=$(pwd)/src:$PYTHONPATH