--- a/backend/src/neuroagent/app/routers/qa.py
+++ b/backend/src/neuroagent/app/routers/qa.py
@@ -144,7 +144,7 @@ async def question_suggestions(
             )
 
     if is_in_chat:
-        content = f"CONVERSATION MESSAGES: \n{json.dumps([{k: v for k, v in json.loads(msg.content).items() if k in ['role', 'content']}for msg in db_messages])}"
+        content = f"CONVERSATION MESSAGES: \n{json.dumps([{k: v for k, v in json.loads(msg.content).items() if k in ['role', 'content']} for msg in db_messages])}"
     else:
         content = (
             f"USER JOURNEY: \n{body.model_dump(exclude={'thread_id'})['click_history']}"