--- a/backend/src/neuroagent/app/stream.py
+++ b/backend/src/neuroagent/app/stream.py
@@ -29,7 +29,9 @@ async def stream_agent_response(
 
     # Restore the httpx client
     httpx_client = AsyncClient(
-        timeout=300.0, verify=False, headers=context_variables["httpx_client"].headers
+        timeout=300.0,
+        verify=False,
+        headers=context_variables["httpx_client"].headers,  # nosec: B501
     )
     context_variables["httpx_client"] = httpx_client
     context_variables["openai_client"] = openai_client