--- a/backend/tests/app/test_dependencies.py
+++ b/backend/tests/app/test_dependencies.py
@@ -81,9 +81,9 @@ def test_get_connection_string_full(monkeypatch, patch_required_env):
 
     settings = Settings()
     result = get_connection_string(settings)
-    assert (
-        result == "http://John:Doe@localhost:5000/test"
-    ), "must return fully formed connection string"
+    assert result == "http://John:Doe@localhost:5000/test", (
+        "must return fully formed connection string"
+    )
 
 
 def test_get_starting_agent(patch_required_env, get_weather_tool):