Closes #357 closes #358 

Introducing `uv`: 48s -> 9s (down to 2s when hitting the cache)
parallelizing mypy and tests: 1m50 -> mypy 52s, unit-tests 59s so 1m to do both since they run in parallel.

Overall it went from ~2m10 to 1m20 since the `CI Frontend` is now the longest running job and takes 1m20.